requirements.txt
tensorflow==2.10.0
numpy==1.21.0
opencv-python==4.6.0.66
pillow==9.0.0
matplotlib==3.5.0
tk==0.1.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.4.0
scikit-learn>=1.0.0
albumentations>=1.3.0
tqdm>=4.65.0
flask==3.0.0 